Update lazyusf2 from Cog and mupen64plus
Upstream the changes I made for Cog, which included blocking
out great portions of the dynarec code with ifdefs. Also
include updates to the RSP HLE code, such as memory bounds
limiting on many arguments throughout. This fixes a major
crash with Conker's Bad Fur Day.

+++ pi/pi_controller.c
@@ -63,7 +63,7 @@ static void dma_pi_write(usf_state_t * state, struct pi_controller* pi)
 {
     unsigned int longueur;
     int i;
-
+    
 #ifdef DEBUG_INFO
     if (state->debug_log)
       fprintf(state->debug_log, "PI DMA WRITE: %08x to %08x for %08x bytes\n", pi->regs[PI_CART_ADDR_REG], pi->regs[PI_DRAM_ADDR_REG], pi->regs[PI_WR_LEN_REG] + 1);
@@ -101,7 +101,7 @@ static void dma_pi_write(usf_state_t * state, struct pi_controller* pi)
     longueur = (pi->regs[PI_WR_LEN_REG] & 0xFFFFFF)+1;
     i = (pi->regs[PI_CART_ADDR_REG]-0x10000000)&0x3FFFFFF;
     longueur = (i + (int) longueur) > pi->cart_rom.rom_size ?
-               (pi->cart_rom.rom_size - i) : longueur;
+               (unsigned int)(pi->cart_rom.rom_size - i) : longueur;
     longueur = (pi->regs[PI_DRAM_ADDR_REG] + longueur) > 0x7FFFFF ?
                (0x7FFFFF - pi->regs[PI_DRAM_ADDR_REG]) : longueur;
 
@@ -132,7 +132,7 @@ static void dma_pi_write(usf_state_t * state, struct pi_controller* pi)
                         bit_array_set(state->barray_ram_written_first, ram_address / 4);
                 }
             }
-
+            
             ((unsigned char*)pi->ri->rdram.dram)[ram_address^S8]=
                 pi->cart_rom.rom[rom_address^S8];
